pyproject.toml
[project]
name = "ozondropbot"
version = "0.1.0"
description = "Telegram bot for tracking Ozon price drops"
requires-python = ">=3.11"
dependencies = [
  "aiogram>=3.4.1",
  "aiosqlite>=0.20.0",
  "apscheduler>=3.10.4",
  "matplotlib>=3.8.0",
  "seaborn>=0.13.0",
  "structlog>=24.1.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
playwright = ["playwright>=1.44.0"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
